backend + frontend
Database fixed, more buttons are now working, bots are still only checking, game is displayed in the ui with dealt cards, player moves and board cards. Still need to fix so that it saves to database and displays it in the ui, fix other buttons such as bet or raise and make it so that bots do other moves rather than just checking.